backend/requirements.txt
fastapi
uvicorn
motor
sumy
pydantic
python-dotenv
transformers
torch
openai
logger
pytest
sentencepiece